+++ backend/src/controllers/Orders/addOrders.js
@@ -1,4 +1,4 @@
-const { Order, Products, Users } = require('../db')
+const { Order, Products, Users } = require('../../db')
 
 const create_order = async (order) => {
     const newOrder = await Orders.create();
