+++ backend/src/controllers/Orders/getOrder.js
@@ -1,4 +1,4 @@
-const { Order, Products, Users } = require('../db');
+const { Order, Products, Users } = require('../../db');
 
 const get_order = async (req, res) => {
     try {
