+++ backend/src/controllers/Product/deleteProducts.js
@@ -24,7 +24,7 @@
 //     delete_Product
 // }
 
-const { Products } = require('../db');
+const { Products } = require('../../db');
 
 const delete_Product = async (req, res) => {
     const { id } = req.params;
