+++ backend/src/controllers/Product/editProduct.js
@@ -1,4 +1,4 @@
-const { Products } = require('../db');
+const { Products } = require('../../db');
 
 const edit_product = async (productId, field, newValue) => {
     try {
@@ -20,7 +20,10 @@ const edit_product = async (productId, field, newValue) => {
 
 const initialEdit = async (req, res) => {
     try {
-        const { productId, field, newValue } = req.body;
+      const {
+        productId,
+        field,
+        newValue } = req.body;
         const productEdit = await edit_product(productId, field, newValue);
         res.status(200).json(productEdit)
     } catch (error) {
