+++ backend/src/controllers/Users/editUser.js
@@ -1,4 +1,4 @@
-const { Users } = require('../db');
+const { Users } = require('../../db');
 
 const edit_user = async (userId, field, newValue) => {
     try {
