+++ backend/src/routes/puts.js
@@ -2,8 +2,8 @@ const { Router } = require('express');
 const fs = require('fs');
 const router_put = Router();
 const multer = require('multer');
-const { initialEdit } = require('../controllers/editProduct');
-const { initialEdit_user } = require('../controllers/editUser');
+const { initialEdit } = require('../controllers/Product/editProduct');
+const { initialEdit_user } = require('../controllers/Users/editUser');
 const upload = multer({ dest: 'uploads/' })
 
 router_put.put('/product', initialEdit); // editar product
